+++ R/imports.R
@@ -6,4 +6,6 @@
 #' @import splines2
 #' @import sandwich
 #' @import ggplot2
+#' @import MASS
+#' @import npiv
 NULL