+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Initial CRAN submission.
 
+* CCK estimation is working for case with two periods.
+
 # contdid 0.0.002
 
 * Added sanity checks for unexpected input data.